fix: Set compiler css option to 'none'
The `css` option to Svelte `compile()` has changed and a `boolean` value is deprecated.

This gave the `compilerOptions.css as a boolean is deprecated. Use 'external' instead of false.` warning.

Since we're not using the, potentially, generated `css` field what so ever, we can set this to `none` and skip all CSS stuff altogether.